requirement.txt
beautifulsoup4==4.9.1
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.9
pafy==0.5.5
PyAudio==0.2.11
python-vlc==3.0.9113
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
urllib3==1.25.9
youtube-dl==2020.6.6
youtube-search==0.1.4
